requirements.txt
streamlit
openai
guardrails-ai